+++ CHANGES.md
@@ -1,3 +1,10 @@
+## 0.10.0 (XXX)
+
+- Disabled credential access boundary by default.
+- Added new base client library: `broker-client-lib`.
+- Renamed `broker-connector` package to `broker-hadoop-connector`.
+- Added `CheckConfig` command.
+
 ## 0.9.2 (March 9, 2020)
 
 - Added support for credentials access boundary.